Fixed layers in PSD and TIFF to show alpha and rgb.
Former-commit-id: a7424a133c02718d5959ed665497a99b17f20909

+++ mrViewer/src/gui/mrvImageBrowser.cpp
@@ -2535,7 +2535,7 @@ void ImageBrowser::handle_dnd()
              if ( ! img->stopped() ) img->stop();
 	     img->seek( f );
 
-             if ( i < children() )
+             if ( (int) i < children() )
                  change_image((int)i);
 
 	  }
